pyproject.toml
[tool.poetry]
name = "02-crew-ai-cold-email-fr"
version = "0.1.0"
description = "Send personalized cold outreach emails to a list of leads."
authors = ["Fanch <fanch.daniel@digi-bo.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0, <3.13"
crewai = {extras = ["tools"], version = "^0.22.5"}
pydantic = "^2.6.3"
load-dotenv = "^0.1.0"
crewai-tools = "^0.0.15"
langchain-groq = "^0.0.1"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"